refactor cxg_admin scripts for clarity/usability
